Major changes!!  This is a BROKEN version be warned... This is the beginning of the major redesign using glade files to define the dynamic portions of the gui.
New so far:  tabloader.{ch}. Calls Glade functions to load the tabs listed
in the interrogation profile, and then binds widget data to the widgets
by reading a config file with the same basename as the glade file but with
a "datamap" extension.  Currently working wiht the B&G 2.0x firmware for the
VE-Table page ONLY.  Other pages to follow.

Currently using a hack to get the widgets from the xml from glade, by using
a PRIVATE header file from libglade.  This may break if glade is different
from 2.0.1. Unfortunately there is no current workaround known.

Added a new dir called "Gui" which contains .glade and .datamap files with
the ame basename.  the .glade file defines the gui layout and structure, the
.datamap contains the widget names and the values to be bound to those
widgets...