Added several flags to the Gui daamap files, alt_depend_on, alt_depend_state, toggle_group. 	"toggle_group" means to toggle the state of the group (string name) when this widget changes state...  Used on dt_params.gui to disable the enhanced idle controls when B&G idle mode is selected... :)
	"alt_depend_on" is used mainly for widgets that have multiple labels
based on some setting,  The most common is temp dependant labels,  so that
there is a fahrenheit label and a celsius label.  The alt_depend_on allows
more label states i.e .for a control that has multiple purposes that DEPENDS
on another controls' state.  It is a string name of the widget it depends on
"alt_depend_state" is the state of the remote widget to determine which set
of labels are valid, the primary or the alternate's ("alt_" prefixed)
the swap_label function uses these widget fields, as does the temp conversions
functions.