Incorporated a "customized" version of libmatheval-1.1.1 from the gnu project.  Library names has been altered from libmatheval to libmtxmatheval to prevent stomping on a potentially already installed instance.  The public header file has also been renamed accordingly.
This library gets installed under /usr/local/lib (default ./configure options)
and megatunix is linked to it.  Probbaly not the best way,  but I couldn't
figure out a clean way to keep it as a shared lib but outside the normal lib
dirs and still make it easily loaded without fiddling with LD_LIBRARY_PATH
or other environ vars,  so the rename method was chosen instead.

This lib is used for storing calculation strings in the config files and
making  them easily parseable,  allowing EXTREME flexibility for complex
numerical conversions for incoming data from the ECU. (can even do
derivatives..)