A big change to the interrogation profiles and tabs themselves...
Details:
  Former methodology,  1 glade file and .datamap per tab, always bound to
eachother.  Problems: If you have multipel tabs that reuse 99% of the controls
you need to create seperate glade files for each one. (wasteful)
 New methodology,  decoupled the glade files from the datamaps a bit, so that
for firmwares hat use multiple similar tables (DT, MSns-Extra, MS-II) only
1 glade file is needed and you jsut select the datamap to set the values
correctlet PER TAB.  So the interrogation profiles have a new line called:
"TabConfs" which is a comma seperated list of files to load respectively
with the "LoadTabs" line. (LoadTabs is hte list of .glade files, Tabconfs
is the list of .datamaps that match up with the glade files.  Thus we can now
load the same glade file multiple times with different configurations for
each load.

New fields added to datamap files, "id_prefix" is a string prepended to the
widget names before storing in the global list of widgets.  Teh other new
field is "set_label" which is ONLY applicable to label widgets, it sets the
text of the label to this text. Required when using a generic glade fiel for
multiple tables when you need to change the title and such.